+++ pyls/config/flake8_conf.py
@@ -1,7 +1,9 @@
 # Copyright 2017 Palantir Technologies, Inc.
 import logging
 import os
-from .source import ConfigSource, parse_config
+from .source import ConfigSource
+from . import _utils
+from pyls._utils import find_parents
 
 log = logging.getLogger(__name__)
 
@@ -21,7 +23,7 @@ class Flake8Config(ConfigSource):
     def user_config(self):
         config_file = self._user_config_file()
         config = self.read_config_from_files([config_file])
-        return parse_config(config, CONFIG_KEY, OPTIONS)
+        return _utils.parse_config(config, CONFIG_KEY, OPTIONS)
 
     def _user_config_file(self):
         if self.is_windows:
@@ -30,7 +32,10 @@ class Flake8Config(ConfigSource):
             return os.path.join(self.xdg_home, 'flake8')
 
     def project_config(self, path=None):
-        # FIXME: this isn't in root_path, this is relative to the file...
-        config_files = [os.path.join(self.root_path, filename) for filename in PROJECT_CONFIGS]
-        config = self.read_config_from_files(config_files)
-        return parse_config(config, CONFIG_KEY, OPTIONS)
+        if path:
+            files = find_parents(self.root_path, path, PROJECT_CONFIGS)
+        else:
+            files = [os.path.join(self.root_path, filename) for filename in PROJECT_CONFIGS]
+
+        config = self.read_config_from_files(files)
+        return _utils.parse_config(config, CONFIG_KEY, OPTIONS)
