+++ pyls/config/pycodestyle_conf.py
@@ -1,7 +1,9 @@
 # Copyright 2017 Palantir Technologies, Inc.
 import os
 import pycodestyle
-from .source import ConfigSource, parse_config
+from .source import ConfigSource
+from . import _utils
+from pyls._utils import find_parents
 
 
 CONFIG_KEY = 'pycodestyle'
@@ -9,7 +11,12 @@ USER_CONFIGS = [pycodestyle.USER_CONFIG] if pycodestyle.USER_CONFIG else []
 PROJECT_CONFIGS = ['pycodestyle.cfg', 'setup.cfg', 'tox.ini']
 
 OPTIONS = [
+    ('exclude', 'plugins.pycodestyle.exclude', list),
+    ('filename', 'plugins.pycodestyle.filename', list),
+    ('hang_closing', 'plugins.pycodestyle.hangClosing', bool),
+    ('ignore', 'plugins.pycodestyle.ignore', list),
     ('max-line-length', 'plugins.pycodestyle.maxLineLength', int),
+    ('select', 'plugins.pycodestyle.select', list),
 ]
 
 
@@ -18,7 +25,7 @@ class PyCodeStyleConfig(ConfigSource):
     def user_config(self):
         config_file = self._user_config_file()
         config = self.read_config_from_files([config_file])
-        return parse_config(config, CONFIG_KEY, OPTIONS)
+        return _utils.parse_config(config, CONFIG_KEY, OPTIONS)
 
     def _user_config_file(self):
         if self.is_windows:
@@ -27,4 +34,10 @@ class PyCodeStyleConfig(ConfigSource):
             return os.path.join(self.xdg_home, 'pycodestyle')
 
     def project_config(self, path=None):
-        return {}
+        if path:
+            files = find_parents(self.root_path, path, PROJECT_CONFIGS)
+        else:
+            files = [os.path.join(self.root_path, filename) for filename in PROJECT_CONFIGS]
+
+        config = self.read_config_from_files(files)
+        return _utils.parse_config(config, CONFIG_KEY, OPTIONS)
