+++ pyls/config/source.py
@@ -4,7 +4,7 @@ import logging
 import os
 import sys
 
-from . import find_parents
+from . import _utils
 
 log = logging.getLogger(__name__)
 
@@ -28,7 +28,7 @@ class ConfigSource(object):
         raise NotImplementedError()
 
     def find_project_files(self, path, filenames):
-        return find_parents(self.root_path, path, filenames)
+        return _utils.find_parents(self.root_path, path, filenames)
 
     def read_config_from_files(self, files):
         # TODO(gatesn): check if files updated
@@ -37,40 +37,3 @@ class ConfigSource(object):
         for filename in files:
             found_files.extend(config.read(filename))
         return config
-
-
-def parse_config(config, key, options):
-    """Parse the config with the given options."""
-    conf = {}
-    for source, destination, opt_type in options:
-        opt_value = _get_opt(config, key, source, opt_type)
-        _set_opt(conf, destination, opt_value)
-    return conf
-
-
-def _get_opt(config, key, option, opt_type):
-    """Get an option from a configparser with the given type."""
-    for opt_key in [option, option.replace('-', '_')]:
-        if not config.has_option(key, opt_key):
-            continue
-
-        if opt_type == int:
-            return config.getint(key, opt_key)
-
-        return config.get(key, opt_key)
-
-
-def _set_opt(config_dict, path, value):
-    """Set the value in the dictionary at the given path if the value is not None."""
-    if value is None:
-        return
-
-    if '.' not in path:
-        config_dict[path] = value
-        return
-
-    key, rest = path.split(".", 1)
-    if key not in config_dict:
-        config_dict[key] = {}
-
-    _set_opt(config_dict[key], rest, value)
