+++ pyls/plugins/pycodestyle_lint.py
@@ -1,7 +1,7 @@
 # Copyright 2017 Palantir Technologies, Inc.
 import logging
 import pycodestyle
-from pyls import config as pyls_config, hookimpl, lsp, _utils
+from pyls import hookimpl, lsp
 
 log = logging.getLogger(__name__)
 
@@ -11,37 +11,19 @@ CONFIG_FILES = ['tox.ini', 'pep8.cfg', 'setup.cfg', 'pycodestyle.cfg']
 
 @hookimpl
 def pyls_lint(config, document):
-    # Start with empty config, and update with each level of settings
-    # conf = {}
+    settings = config.plugin_settings('pycodestyle')
+    log.debug("Got pycodestyle settings: %s", settings)
 
-    # def _config_from_files(config_files):
-    #     pycodestyle_conf = pyls_config.build_config('pycodestyle', config_files)
-    #     pep8_conf = pyls_config.build_config('pep8', config_files)
-    #     conf_to_use = pycodestyle_conf or pep8_conf or {}
-    #     return {k.replace("-", "_"): v for k, v in conf_to_use.items()}
+    opts = {
+        'exclude': settings.get('exclude'),
+        'filename': settings.get('filename'),
+        'hang_closing': settings.get('hangClosing'),
+        'ignore': settings.get('ignore'),
+        'max_line_length': settings.get('maxLineLength'),
+        'select': settings.get('select'),
+    }
 
-    # # First, read the user configuration (in user's home directory)
-    # if pycodestyle.USER_CONFIG:
-    #     conf.update(_config_from_files([pycodestyle.USER_CONFIG]))
-
-    # # Then, read the PYLS configuration
-    # ide_conf = config.plugin_settings('pycodestyle')
-    # conf.update({_utils.camel_to_underscore(k): _utils.list_to_string(v) for k, v in ide_conf.items()})
-
-    # # Finally, read the project configuration
-    # conf.update(_config_from_files(config.find_parents(document.path, CONFIG_FILES)))
-
-    # Grab the pycodestyle parser and set the defaults based on the config we found
-
-    conf = {}
-
-    parser = pycodestyle.get_parser()
-    parser.set_defaults(**conf)
-
-    opts, _args = parser.parse_args([])
-    log.debug("Got pycodestyle config: %s", opts)
-    styleguide = pycodestyle.StyleGuide(vars(opts))
-
+    styleguide = pycodestyle.StyleGuide({k: v for k, v in opts.items() if v is not None})
     c = pycodestyle.Checker(
         filename=document.uri, lines=document.lines, options=styleguide.options,
         report=PyCodeStyleDiagnosticReport(styleguide.options)
@@ -53,9 +35,9 @@ def pyls_lint(config, document):
 
 class PyCodeStyleDiagnosticReport(pycodestyle.BaseReport):
 
-    def __init__(self, options=None, **kwargs):
+    def __init__(self, options=None):
         self.diagnostics = []
-        super(PyCodeStyleDiagnosticReport, self).__init__(options=options, **kwargs)
+        super(PyCodeStyleDiagnosticReport, self).__init__(options=options)
 
     def error(self, lineno, offset, text, check):
         # PyCodeStyle will sometimes give you an error the line after the end of the file
