+++ test/plugins/test_lint.py
@@ -1,7 +1,7 @@
 # Copyright 2017 Palantir Technologies, Inc.
 import os
 from pyls import lsp, uris
-from pyls.config import Config
+from pyls.config.config import Config
 from pyls.workspace import Document
 from pyls.plugins import mccabe_lint, pycodestyle_lint, pydocstyle_lint, pyflakes_lint
